Add VTK_GEO_XML importers
Note: for the moment the importers and the exporters are on different
interfaces. In particular it has been chosen to distinguigh them with
the suffix "_WRITE" and "_READ" for exporters and importers,
respectively. This break the BACKWARD COMPATIBILITY. In the future OOP
refactoring it will investigated how to give a simple generic interface
for both importers and exporters.