Add the skeleton for binary local volatile writer
Add the skeleton for binary local volatile writer: a writer that saves
data into a binary local volatile characters string that can be easily
shared between different processes and can be created also by processes
that have not access to the filesystem for creating a real file.

Just a skeleton not even compilable.